requirements.txt
protobuf==3.20.1
pandas
plotly
streamlit
openpyxl
streamlit-extras
google-api-python-client
requests
google-auth-oauthlib
google-auth-httplib2
streamlit-authenticator==0.1.5
